Added the movie playlist system
Created the movie playlist system, the playlists are defined in the
movies_setters.js file as functions and one needs to be created per
playlist. The system is also setup to hide the playlist mechanism when a
singular movie is selected.
Current movie playlist is a demo and does not have correct files.